Removed All Suspicious BS Activity Crashing Code
Couldn't figure out what was causing the crash when the user attempts to navigate to the Ballisitic skill page. Worked fine without any of the new code added to BSActivity.class. 
Ended up just deleting that one class and gonna see how to write it. Maybe next time it will work as intended