New Inventory Layout
Had to rework the tab2Inventory layout file to be able to parse die (xdy+z) & and RoF (x/y/z) info easily. Cleaned up the layout formatting too to use linear layouts to provide a simpler (drag and drop) function for later development.

Depending on the scope of this project. This may one day be turned into a recyclerView because I am pretty heavily using too many views on that tab and it /may/ one day affect speeds